+++ unity/io.github.xoderony.channels/Runtime/ValueChannels/IValueReader.cs
@@ -1,0 +1,6 @@
+namespace Xoderony;
+
+public interface IValueReader<T> {
+
+    ref readonly T? Value { get; }
+}