+++ unity/io.github.xoderony.core/Runtime/ValueChannels/IValueReader.cs
@@ -1,7 +1,0 @@
-namespace Xoderony;
-
-public interface IValueReader<T> {
-
-    ref readonly T? Value { get; }
-
-}